mount marketing app into container
We use the mount function to mount the marketing app into the container instead of exporting component from marketing app and rendering it directly. This way we decouple the container and marketing. If someday, container uses some other framework then we just need to call mount from marketing according to new framework and doesn't need to change anything in marketing app.